feat(wolves): render comic pages on demand
- Add WolvesComicReader.vue replacing partial ComicReader.vue
  - Accepts chapters: WolvesChapter[] prop; derives PDF URL internally
  - Defaults to 'paged' mode (not scroll)
  - Mode controls use role=tablist / role=tab / aria-selected
  - Navigation buttons: aria-label='Previous page' / 'Next page', disabled binding
  - Emits update:page (1-based) and chapterChange on boundary cross
  - Continuous mode uses IntersectionObserver (rootMargin: 800px 0px)
    to render approaching pages on demand — no renderAllScrollPages loop
  - Observer and renderTasks cleaned up on unmount
- Add wolvesComicReader.test.ts — 9 tests covering paged default,
  tablist/aria-selected, update:page emission, chapter boundary, cleanup
- Update WolvesApp.vue to compose with WolvesComicReader:
    <WolvesComicReader :chapters='wolvesRelease.chapters' @update:page='currentPage = $event' />
  - currentPage ref is 1-based; activeChapter derives from currentPage directly
  - Remove PDF_URL const and lore-prev/lore-next bindings from reader
