+++ src/WolvesApp.vue
@@ -2,11 +2,8 @@
 README: Bluefin Wolves Teaser Landing Page Component
 ===================================================
 - Page Path: projectbluefin.io/wolves
-- Comic Content: Renders the real "Color with Bluefin" coloring book PDF
-  (https://download.projectbluefin.io/color-with-bluefin.pdf) page-by-page onto
-  an HTML5 canvas using PDF.js, loaded dynamically from the cdnjs CDN (see
-  `PDFJS_SCRIPT_URL` / `PDFJS_WORKER_URL` below). To point at a different PDF,
-  update `PDF_URL`.
+- Comic Content: Delegated to WolvesComicReader.vue. PDF URL is managed
+  by the component itself (BASE_URL + color-with-bluefin.pdf).
 - Intercepted Communications: Sourced from `src/data/intercepted-communications.json`.
   Add conversations there with title, channel, date, and ordered messages.
 - Donate QR Code: Pointing to `https://docs.projectbluefin.io/donations`.
@@ -15,13 +12,15 @@ README: Bluefin Wolves Teaser Landing Page Component
 -->
 <script setup lang="ts">
 import type { WolvesChapter } from './data/wolves-story'
-import { computed, nextTick, onBeforeUnmount, onMounted, ref, watch } from 'vue'
+import { computed, onBeforeUnmount, onMounted, ref } from 'vue'
 import qrDonate from '@/assets/svg/qr-donate.svg'
 import qrStore from '@/assets/svg/qr-store.svg'
 import TopNavbar from './components/TopNavbar.vue'
+import WolvesComicReader from './components/wolves/WolvesComicReader.vue'
 import WolvesSoundtrack from './components/wolves/WolvesSoundtrack.vue'
 import bazziteQuotes from './data/bazzite-quotes.json'
 import interceptedCommunications from './data/intercepted-communications.json'
+import { wolvesRelease } from './data/wolves-story'
 import { shuffleLoreEntries } from './utils/loreRotation'
 import { getChapterForPage } from './utils/wolvesStory'
 
@@ -63,268 +62,12 @@ const loreEntries = shuffleLoreEntries([
   ...conversations.map(data => ({ type: 'conversation' as const, data })),
 ])
 
-// PDF.js is injected dynamically from CDNJS (see loadPdfJs()) and attaches itself
-// to `window.pdfjsLib`. It ships no first-party types, so we treat it as `any`.
-const PDF_URL = `${import.meta.env.BASE_URL}color-with-bluefin.pdf`
-const PDFJS_SCRIPT_URL = 'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.11.174/pdf.min.js'
-const PDFJS_SCRIPT_INTEGRITY = 'sha512-q+4liFwdPC/bNdhUpZx6aXDx/h77yEQtn4I1slHydcbZK34nLaR3cAeYSJshoxIOq3mjEf7xJE8YWIUHMn+oCQ=='
-const PDFJS_WORKER_URL = 'https://cdnjs.cloudflare.com/ajax/libs/pdf.js/3.11.174/pdf.worker.min.js'
-
-let pdfjsLib: any = null
-let pdfDocument: any = null
-// Tracks in-flight render tasks per canvas element so a resize/page-flip can cancel
-// a stale render before starting a new one on the same canvas.
-const renderTasks = new Map<HTMLCanvasElement, any>()
-
-// Comic Reader state
-const totalPages = ref(0)
-const currentPageIndex = ref(0) // 0-based for the UI; PDF.js pages are 1-based
-const readingMode = ref<'flip' | 'scroll'>('scroll') // 'flip' = page-by-page, 'scroll' = stacked vertical
-const pdfLoading = ref(true)
-const pdfError = ref('')
-
 // Soundtrack / entry state (component-owned detail lives in WolvesSoundtrack)
 const hasEntered = ref(false)
-const activeChapter = computed<WolvesChapter | undefined>(() => getChapterForPage(currentPageIndex.value + 1))
 
-// Template refs: the "flip" mode uses a single canvas, the "scroll" mode renders
-// one canvas per page into an array indexed by page number.
-const flipViewport = ref<HTMLElement | null>(null)
-const flipCanvas = ref<HTMLCanvasElement | null>(null)
-const scrollContainer = ref<HTMLElement | null>(null)
-const scrollCanvases = ref<(HTMLCanvasElement | null)[]>([])
-
-let flipResizeObserver: ResizeObserver | null = null
-let scrollResizeObserver: ResizeObserver | null = null
-
-function setScrollCanvasRef(el: Element | null, index: number) {
-  scrollCanvases.value[index] = el as HTMLCanvasElement | null
-}
-
-// Dynamically injects a <script> tag once and resolves when it has loaded.
-// When `integrity` is supplied, the script is fetched with subresource integrity
-// verification and CORS mode enabled so the browser can check the hash.
-function loadScript(src: string, integrity?: string): Promise<void> {
-  return new Promise((resolve, reject) => {
-    if (document.querySelector(`script[src="${src}"]`)) {
-      resolve()
-      return
-    }
-    const script = document.createElement('script')
-    script.src = src
-    script.async = true
-    if (integrity) {
-      script.integrity = integrity
-      script.crossOrigin = 'anonymous'
-    }
-    script.onload = () => resolve()
-    script.onerror = () => reject(new Error(`Failed to load script: ${src}`))
-    document.head.appendChild(script)
-  })
-}
-
-// Injects PDF.js from CDNJS and wires up its worker script, caching the global for reuse.
-async function loadPdfJs(): Promise<any> {
-  if (pdfjsLib) {
-    return pdfjsLib
-  }
-  await loadScript(PDFJS_SCRIPT_URL, PDFJS_SCRIPT_INTEGRITY)
-  const lib = (window as any).pdfjsLib
-  if (!lib) {
-    throw new Error('PDF.js failed to initialize')
-  }
-  lib.GlobalWorkerOptions.workerSrc = PDFJS_WORKER_URL
-  pdfjsLib = lib
-  return lib
-}
-
-// Renders a single PDF page onto the given canvas, scaled to fill containerWidth.
-async function renderPageOnCanvas(pageNumber: number, canvas: HTMLCanvasElement, containerWidth: number) {
-  if (!pdfDocument || containerWidth <= 0) {
-    return
-  }
-
-  // Cancel a stale in-flight render targeting this canvas before starting a new one.
-  const existingTask = renderTasks.get(canvas)
-  if (existingTask) {
-    existingTask.cancel()
-  }
-
-  const page = await pdfDocument.getPage(pageNumber)
-  const baseViewport = page.getViewport({ scale: 1 })
-  const scale = containerWidth / baseViewport.width
-  const viewport = page.getViewport({ scale })
-
-  const context = canvas.getContext('2d')
-  if (!context) {
-    return
-  }
-
-  // Render at native device pixel ratio for crisp output on HiDPI screens.
-  const outputScale = window.devicePixelRatio || 1
-  canvas.width = Math.floor(viewport.width * outputScale)
-  canvas.height = Math.floor(viewport.height * outputScale)
-  canvas.style.width = `${Math.floor(viewport.width)}px`
-  canvas.style.height = `${Math.floor(viewport.height)}px`
-  const transform = outputScale !== 1 ? [outputScale, 0, 0, outputScale, 0, 0] : undefined
-
-  const renderTask = page.render({ canvasContext: context, viewport, transform })
-  renderTasks.set(canvas, renderTask)
-  renderTask.promise.then(() => {
-    if (renderTasks.get(canvas) === renderTask) {
-      renderTasks.delete(canvas)
-    }
-  }).catch(() => {})
-  try {
-    await renderTask.promise
-  }
-  catch (err: any) {
-    if (err?.name !== 'RenderingCancelledException') {
-      console.error('[wolves] Failed to render PDF page', pageNumber, err)
-    }
-  }
-}
-
-function getContentWidth(element: HTMLElement): number {
-  const styles = window.getComputedStyle(element)
-  const paddingLeft = Number.parseFloat(styles.paddingLeft) || 0
-  const paddingRight = Number.parseFloat(styles.paddingRight) || 0
-  return Math.max(0, element.clientWidth - paddingLeft - paddingRight)
-}
-
-function renderFlipPage() {
-  if (!flipCanvas.value) {
-    return
-  }
-  const host = flipCanvas.value.parentElement as HTMLElement | null
-  const width = host ? getContentWidth(host) : flipViewport.value?.clientWidth ?? 0
-  renderPageOnCanvas(currentPageIndex.value + 1, flipCanvas.value, width)
-}
-
-function renderAllScrollPages() {
-  if (!scrollContainer.value) {
-    return
-  }
-  for (let i = 0; i < totalPages.value; i++) {
-    const canvas = scrollCanvases.value[i]
-    if (canvas) {
-      const host = canvas.parentElement as HTMLElement | null
-      const width = host ? getContentWidth(host) : scrollContainer.value.clientWidth
-      renderPageOnCanvas(i + 1, canvas, width)
-    }
-  }
-}
-
-function setupFlipResizeObserver() {
-  flipResizeObserver?.disconnect()
-  if (!flipViewport.value) {
-    return
-  }
-  flipResizeObserver = new ResizeObserver(() => renderFlipPage())
-  flipResizeObserver.observe(flipViewport.value)
-}
-
-function setupScrollResizeObserver() {
-  scrollResizeObserver?.disconnect()
-  if (!scrollContainer.value) {
-    return
-  }
-  scrollResizeObserver = new ResizeObserver(() => renderAllScrollPages())
-  scrollResizeObserver.observe(scrollContainer.value)
-}
-
-async function loadComicPdf() {
-  pdfLoading.value = true
-  pdfError.value = ''
-  try {
-    const lib = await loadPdfJs()
-    pdfDocument = await lib.getDocument(PDF_URL).promise
-    totalPages.value = pdfDocument.numPages
-    pdfLoading.value = false
-    await nextTick()
-    if (readingMode.value === 'flip') {
-      setupFlipResizeObserver()
-      renderFlipPage()
-    }
-    else {
-      setupScrollResizeObserver()
-      renderAllScrollPages()
-    }
-  }
-  catch (err) {
-    console.error('[wolves] Failed to load comic PDF', err)
-    pdfError.value = 'Unable to load the comic book right now. Please try again in a moment.'
-    pdfLoading.value = false
-  }
-}
-
-// Keyboard navigation helper
-function handleKeyDown(event: KeyboardEvent) {
-  const target = event.target as HTMLElement
-  if (target && (['INPUT', 'SELECT', 'TEXTAREA'].includes(target.tagName) || target.isContentEditable)) {
-    return
-  }
-
-  if (event.key === 'ArrowRight' || event.key === 'Right') {
-    if (readingMode.value === 'flip') {
-      nextPage()
-    }
-    else {
-      loreNext()
-    }
-  }
-  else if (event.key === 'ArrowLeft' || event.key === 'Left') {
-    if (readingMode.value === 'flip') {
-      prevPage()
-    }
-    else {
-      lorePrev()
-    }
-  }
-}
-
-function nextPage() {
-  if (currentPageIndex.value < totalPages.value - 1) {
-    currentPageIndex.value++
-  }
-}
-
-function prevPage() {
-  if (currentPageIndex.value > 0) {
-    currentPageIndex.value--
-  }
-}
-
-function jumpToPage(index: number) {
-  if (index >= 0 && index < totalPages.value) {
-    currentPageIndex.value = index
-  }
-}
-
-// Re-render the current page whenever it changes in flip mode.
-watch(currentPageIndex, () => {
-  if (readingMode.value === 'flip' && !pdfLoading.value) {
-    renderFlipPage()
-  }
-})
-
-// Swap resize observers and (re)render pages whenever the layout mode toggles.
-watch(readingMode, async (mode) => {
-  if (pdfLoading.value || pdfError.value) {
-    return
-  }
-  await nextTick()
-  if (mode === 'flip') {
-    scrollResizeObserver?.disconnect()
-    setupFlipResizeObserver()
-    renderFlipPage()
-  }
-  else {
-    flipResizeObserver?.disconnect()
-    setupScrollResizeObserver()
-    renderAllScrollPages()
-  }
-})
+// Current page (1-based) tracked here so the chapter can be passed to WolvesSoundtrack.
+const currentPage = ref(1)
+const activeChapter = computed<WolvesChapter | undefined>(() => getChapterForPage(currentPage.value))
 
 // Mixed lore cycling state. The source arrays are shuffled once per page load.
 const currentLoreIndex = ref(0)
@@ -369,22 +112,10 @@ function lorePrev() {
 }
 
 onMounted(() => {
-  window.addEventListener('keydown', handleKeyDown)
-  loadComicPdf()
-
-  // Start mixed lore auto-cycling interval (9 seconds)
   startLoreTimer()
 })
 
 onBeforeUnmount(() => {
-  window.removeEventListener('keydown', handleKeyDown)
-  flipResizeObserver?.disconnect()
-  scrollResizeObserver?.disconnect()
-  renderTasks.forEach(task => task.cancel())
-  renderTasks.clear()
-  pdfDocument?.destroy()
-
-  // Clear mixed lore auto-cycling interval to prevent memory leaks
   stopLoreTimer()
 })
 </script>
@@ -428,146 +159,10 @@ onBeforeUnmount(() => {
       <div class="content-grid">
         <div class="col-left">
           <!-- SECTION 2: COMIC READER -->
-          <section id="comic-reader" class="comic-reader-section">
-            <div class="comic-toolbar">
-              <div class="mode-selectors">
-                <button
-                  :class="{ active: readingMode === 'flip' }"
-                  @click="readingMode = 'flip'"
-                >
-                  Page By Page
-                </button>
-                <button
-                  :class="{ active: readingMode === 'scroll' }"
-                  @click="readingMode = 'scroll'"
-                >
-                  Continuous Scroll
-                </button>
-              </div>
-            </div>
-
-            <!-- Comic Reader Layout: Page by Page (Slideshow) -->
-            <div v-if="readingMode === 'flip'" class="page-flip-comic-layout">
-              <div ref="flipViewport" class="comic-viewport">
-                <!-- Page Contents -->
-                <div class="comic-content-area">
-                  <div v-if="pdfLoading" class="comic-status-wrap">
-                    <div class="spinner" />
-                    <p>Loading comic pages&hellip;</p>
-                  </div>
-                  <div v-else-if="pdfError" class="comic-status-wrap is-error">
-                    <p>{{ pdfError }}</p>
-                    <button class="ctrl-btn" @click="loadComicPdf">
-                      Retry
-                    </button>
-                  </div>
-                  <canvas
-                    v-show="!pdfLoading && !pdfError"
-                    ref="flipCanvas"
-                    class="pdf-page-canvas"
-                    role="img"
-                    :aria-label="`Comic page ${currentPageIndex + 1} of ${totalPages}`"
-                  />
-                </div>
-
-                <!-- Left Navigation Button -->
-                <button
-                  v-show="!pdfLoading && !pdfError && currentPageIndex > 0"
-                  class="nav-btn prev"
-                  aria-label="Previous Page"
-                  @click="prevPage"
-                >
-                  &larr;
-                </button>
-
-                <!-- Right Navigation Button -->
-                <button
-                  v-show="!pdfLoading && !pdfError && currentPageIndex < totalPages - 1"
-                  class="nav-btn next"
-                  aria-label="Next Page"
-                  @click="nextPage"
-                >
-                  &rarr;
-                </button>
-              </div>
-
-              <!-- Bottom Control Bar (Navigation Controls) -->
-              <div class="reader-controls">
-                <button
-                  class="ctrl-btn"
-                  :disabled="pdfLoading || !!pdfError || currentPageIndex === 0"
-                  @click="prevPage"
-                >
-                  &larr; Previous
-                </button>
-
-                <!-- Keyboard helper -->
-                <div class="kbd-hint">
-                  Use &larr; &rarr; arrow keys to turn pages &middot; Page {{ currentPageIndex + 1 }} of {{ totalPages || '—' }}
-                </div>
-
-                <div class="jump-select-wrap">
-                  <span>Jump to:</span>
-                  <select
-                    :value="currentPageIndex"
-                    :disabled="pdfLoading || !!pdfError || !totalPages"
-                    @change="jumpToPage(Number(($event.target as HTMLSelectElement).value))"
-                  >
-                    <option v-for="n in totalPages" :key="n" :value="n - 1">
-                      Page {{ n }}
-                    </option>
-                  </select>
-                </div>
-
-                <button
-                  class="ctrl-btn"
-                  :disabled="pdfLoading || !!pdfError || currentPageIndex === totalPages - 1"
-                  @click="nextPage"
-                >
-                  Next &rarr;
-                </button>
-              </div>
-            </div>
-
-            <!-- Comic Reader Layout: Continuous Stacked Vertical Scroll -->
-            <div v-else ref="scrollContainer" class="scroll-comic-layout">
-              <template v-if="pdfLoading">
-                <div class="comic-status-wrap">
-                  <div class="spinner" />
-                  <p>Loading comic pages&hellip;</p>
-                </div>
-              </template>
-              <template v-else-if="pdfError">
-                <div class="comic-status-wrap is-error">
-                  <p>{{ pdfError }}</p>
-                  <button class="ctrl-btn" @click="loadComicPdf">
-                    Retry
-                  </button>
-                </div>
-              </template>
-              <template v-else>
-                <div
-                  v-for="n in totalPages"
-                  :key="n"
-                  class="scroll-page-card"
-                >
-                  <div class="comic-viewport">
-                    <div class="comic-content-area">
-                      <canvas
-                        :ref="(el) => setScrollCanvasRef(el as Element | null, n - 1)"
-                        class="pdf-page-canvas"
-                        role="img"
-                        :aria-label="`Page ${n} of ${totalPages}`"
-                      />
-                    </div>
-                  </div>
-                  <div class="comic-caption-bar">
-                    Page {{ n }} of {{ totalPages }}
-                  </div>
-                </div>
-              </template>
-            </div>
-          </section>
+          <WolvesComicReader
+            :chapters="wolvesRelease.chapters"
+            @update:page="currentPage = $event"
+          />
         </div>
 
         <div class="col-right">
@@ -798,9 +393,9 @@ onBeforeUnmount(() => {
 
   // Left-align the comic reader within its column instead of the
   // page-wide auto-centering used when the reader is the sole column.
-  .comic-viewport,
-  .scroll-comic-layout,
-  .reader-controls {
+  :deep(.comic-viewport),
+  :deep(.scroll-comic-layout),
+  :deep(.reader-controls) {
     margin-left: 0;
     margin-right: 0;
     max-width: 100%;
@@ -891,255 +486,6 @@ onBeforeUnmount(() => {
   100% {
     transform: scale(1);
     opacity: 0.8;
-  }
-}
-
-// Comic Reader Section
-.comic-toolbar {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 12px;
-
-  .comic-hint {
-    margin: 0;
-    font-size: 1.1rem;
-    color: rgba(189, 189, 189, 0.8);
-    text-align: center;
-  }
-}
-
-.mode-selectors {
-  display: flex;
-  background-color: #10151f;
-  padding: 4px;
-  border-radius: 8px;
-  border: 1px solid #272727;
-  align-self: flex-start;
-
-  button {
-    background: none;
-    border: none;
-    color: #bdbdbd;
-    font-size: 1.2rem;
-    font-weight: 700;
-    padding: 6px 12px;
-    border-radius: 6px;
-    cursor: pointer;
-    transition: all 0.2s;
-
-    &.active {
-      background-color: var(--color-blue);
-      color: #ffffff;
-    }
-
-    &:hover:not(.active) {
-      color: #ffffff;
-    }
-  }
-}
-
-// Comic Page Viewer
-.comic-viewport {
-  position: relative;
-  width: 100%;
-  min-height: 220px;
-  max-width: 760px;
-  max-height: min(74dvh, 760px);
-  margin: 0 auto;
-  background-color: #10151f;
-  border: 1px solid rgba(var(--color-blue-rgb), 0.3);
-  border-radius: 16px;
-  overflow: hidden;
-  box-shadow: 0 30px 60px rgba(0, 0, 0, 0.6);
-  display: flex;
-  flex-direction: column;
-
-  .comic-content-area {
-    flex: 1;
-    position: relative;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    min-height: 220px;
-    padding: 12px;
-    overflow: hidden;
-  }
-
-  .nav-btn {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    background-color: rgba(0, 0, 0, 0.7);
-    border: 1px solid #272727;
-    color: #ffffff;
-    font-size: 1.6rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    transition: all 0.2s;
-    z-index: 10;
-
-    &:hover {
-      background-color: #000;
-      color: var(--color-blue-light);
-      border-color: var(--color-blue-light);
-    }
-
-    &.prev {
-      left: 12px;
-    }
-    &.next {
-      right: 12px;
-    }
-  }
-}
-
-// Canvas the current PDF page is rendered onto; JS sets explicit pixel
-// dimensions to match the container width at the device pixel ratio.
-.pdf-page-canvas {
-  display: block;
-  width: auto;
-  height: auto;
-  max-width: 100%;
-  max-height: 100%;
-}
-
-// Loading / error states shown while PDF.js fetches and parses the PDF
-.comic-status-wrap {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 16px;
-  padding: 48px 32px;
-  text-align: center;
-  color: #bdbdbd;
-  font-size: 1.4rem;
-
-  &.is-error {
-    color: var(--color-blue-light);
-  }
-
-  .spinner {
-    width: 36px;
-    height: 36px;
-    border-radius: 50%;
-    border: 3px solid rgba(var(--color-blue-rgb), 0.25);
-    border-top-color: var(--color-blue);
-    animation: comic-spinner-spin 0.8s linear infinite;
-  }
-}
-
-@keyframes comic-spinner-spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.comic-caption-bar {
-  background-color: rgba(0, 0, 0, 0.9);
-  padding: 16px 24px;
-  border-top: 1px solid #272727;
-  text-align: center;
-  font-size: 1.3rem;
-  color: #ffffff;
-  font-weight: 500;
-  line-height: 1.5;
-}
-
-.reader-controls {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  max-width: 760px;
-  margin: 0 auto;
-  padding: 0;
-  gap: 8px 12px;
-  flex-wrap: wrap;
-
-  .ctrl-btn {
-    background-color: #10151f;
-    border: 1px solid #272727;
-    color: #bdbdbd;
-    font-size: 1.2rem;
-    font-weight: 700;
-    padding: 8px 16px;
-    border-radius: 8px;
-    cursor: pointer;
-    transition: all 0.2s;
-
-    &:hover:not(:disabled) {
-      color: #ffffff;
-      border-color: var(--color-blue-light);
-    }
-
-    &:disabled {
-      opacity: 0.3;
-      cursor: not-allowed;
-    }
-  }
-
-  .kbd-hint {
-    font-size: 1.1rem;
-    color: #616161;
-  }
-
-  .jump-select-wrap {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    font-size: 1.2rem;
-    color: #bdbdbd;
-
-    select {
-      background-color: #10151f;
-      border: 1px solid #272727;
-      color: #ffffff;
-      padding: 6px 12px;
-      border-radius: 6px;
-      font-size: 1.2rem;
-      cursor: pointer;
-
-      &:focus {
-        outline: none;
-        border-color: var(--color-blue-light);
-      }
-    }
-  }
-}
-
-// Continuous Scroll Layout
-.scroll-comic-layout {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  width: 100%;
-  max-width: 760px;
-  margin: 0 auto;
-
-  .comic-viewport {
-    max-height: none;
-  }
-
-  .comic-content-area {
-    overflow: visible;
-  }
-
-  .scroll-page-card {
-    background-color: #10151f;
-    border: 1px solid rgba(var(--color-blue-rgb), 0.2);
-    border-radius: 16px;
-    overflow: hidden;
-    box-shadow: 0 10px 30px rgba(0, 0, 0, 0.4);
-    display: flex;
-    flex-direction: column;
   }
 }
